gastarme/requirements.txt
Django==2.0
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0